services/users/requirements.txt
fastapi
uvicorn
google-cloud-firestore
firebase-admin
httpx
pydantic[email]
